+++ core/src/main/scala/tech/beshu/ror/configuration/RorConfig.scala
@@ -17,11 +17,11 @@
 package tech.beshu.ror.configuration
 
 import tech.beshu.ror.RequestId
+import tech.beshu.ror.accesscontrol.audit.AuditingTool
 import tech.beshu.ror.accesscontrol.blocks.ImpersonationWarning
 import tech.beshu.ror.accesscontrol.blocks.definitions.ldap.LdapService
 import tech.beshu.ror.accesscontrol.blocks.definitions.{ExternalAuthenticationService, ExternalAuthorizationService}
 import tech.beshu.ror.accesscontrol.domain.LocalUsers
-import tech.beshu.ror.accesscontrol.logging.audit.AuditingTool
 import tech.beshu.ror.configuration.RorConfig.ImpersonationWarningsReader
 
 final case class RorConfig(services: RorConfig.Services,
