+++ core/src/test/resources/log4j2.properties
@@ -36,7 +36,8 @@ rootLogger.appenderRef.console.ref = console
 
 appender.readonlyrest_audit_rolling.type = RollingFile
 appender.readonlyrest_audit_rolling.name = readonlyrest_audit_rolling
-appender.readonlyrest_audit_rolling.fileName = ./core/build/tmp/readonlyrest_audit.log
+appender.readonlyrest_audit_rolling.fileName = ${sys:ROR_AUDIT_LOG_FILE_PATH}
+appender.readonlyrest_audit_rolling.createOnDemand = true
 appender.readonlyrest_audit_rolling.layout.type = PatternLayout
 appender.readonlyrest_audit_rolling.layout.pattern = [%d{ISO8601}] %m%n
 appender.readonlyrest_audit_rolling.filePattern = readonlyrest_audit-%i.log.gz
