+++ integration-tests/src/test/resources/cluster_auditing_tools/readonlyrest.yml
@@ -8,9 +8,11 @@ readonlyrest:
   prompt_for_basic_auth: false
 
   audit:
-    collector: true
-    index_template: "'audit_index'"
-    cluster: ["http://{AUDIT_1_HOST}:{AUDIT_1_PORT}"]
+    enabled: true
+    outputs:
+      - type: index
+        index_template: "'audit_index'"
+        cluster: ["http://{AUDIT_1_HOST}:{AUDIT_1_PORT}"]
 
   access_control_rules:
