+++ integration-tests/src/test/resources/enabled_auditing_tools/readonlyrest.yml
@@ -8,8 +8,10 @@ readonlyrest:
   prompt_for_basic_auth: false
 
   audit:
-    collector: true
-    index_template: "'audit_index'"
+    enabled: true
+    outputs:
+      - type: index
+        index_template: "'audit_index'"
 
   access_control_rules:
